lesson2 task2

+++ less2/task1.php
@@ -12,9 +12,9 @@ while (!$isAnswerCorrect) {
     echo "3) 1896 - 1905" . PHP_EOL;
     echo "4) 1896 - 1918" . PHP_EOL;
 
-    $answer = readline("Выберите нужную цифру? ");
+    $answer = (int)readline("Выберите нужную цифру? ");
 
-    if ((int)$answer === $correctAnswerIs) {
+    if ($answer === $correctAnswerIs) {
         echo PHP_EOL . "Ответ верный! Поздравляю!" . PHP_EOL;
         $isAnswerCorrect = true;
     } else {
